// src/types/next-auth.d.ts
import { useValidator } from './../hooks/useValidator';
import NextAuth from "next-auth"
import { JWT } from "next-auth/jwt"

declare module "next-auth" {
    /**
     * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
     */
    interface Session {
        user: {
            id: string;
            name: string;
            authToken: string;
            authTokenExpiration: number;
        },
        authToken: string;
    }
    interface User {
        id: string;
        name: string;
        authToken: string;
        authTokenExpiration: number;
        refreshToken: string;
        validated: boolean;
    }
}

declare module "next-auth/jwt" {
    /** Returned by the `jwt` callback and `getToken`, when using JWT sessions */
    interface JWT {
        id: string;
        name: string;
        authToken: string;
        authTokenExpiration: number;
        refreshToken: string;
    }
}

export interface AuthCredentials {
    username: string;
    password: string;
    rememberMe: string;
}